```
from asv_bench.benchmarks.join_merge import *
self6 = Merge()
self6.setup(False)

%timeit self6.time_merge_dataframes_cross(False)
535 ms ± 4.25 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
242 ms ± 5.98 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- PR


self6.setup(True)
%timeit self6.time_merge_dataframes_cross(True)
540 ms ± 9.13 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- master
243 ms ± 4.57 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)  # <- PR
```